Move normalize_and_snakecase_keys to identifier_utils
