+++ f/common_logic/tests/identifier_utils_test.py
@@ -1,4 +1,8 @@
-from f.common_logic.identifier_utils import camel_to_snake, sanitize_sql_message
+from f.common_logic.identifier_utils import (
+    camel_to_snake,
+    normalize_and_snakecase_keys,
+    sanitize_sql_message,
+)
 
 
 def test_sanitize_sql_message():
@@ -90,3 +94,39 @@ def test_camel_to_snake():
     assert camel_to_snake("camel_case_") == "camel_case_"
     assert camel_to_snake("URLLink") == "url_link"
     assert camel_to_snake("URL_Link") == "url_link"
+
+
+def test_normalize_and_snakecase_keys():
+    input_dict = {
+        "primaryKey": 1,
+        "camelCaseKey": 2,
+        "anotherCamelCaseKey": 3,
+        "keyWith-Collision": 4,
+        "keyWithCollision": 5,
+        "KeyWithCollision": 6,
+        "key-with-collision": 7,
+        "key_with_collision": 8,
+        "key_with_collision_2": 9,
+        "aVeryLongKeyNameThatExceedsTheSixtyThreeCharacterLimitAndNeedsTruncation": 10,
+        "aVeryLongKeyNameThatExceedsTheSixtyThreeCharacterLimitAndNeedsTruncationAlso": 11,
+    }
+
+    special_case_keys = set(["primaryKey"])
+
+    expected_output = {
+        "primaryKey": 1,
+        "camel_case_key": 2,
+        "another_camel_case_key": 3,
+        "key_with_collision": 4,
+        "key_with_collision_2": 5,
+        "key_with_collision_3": 6,
+        "key_with_collision_4": 7,
+        "key_with_collision_5": 8,
+        "key_with_collision_2_2": 9,
+        "a_very_long_key_name_that_exceeds_the_sixty_three_character_l_1": 10,
+        "a_very_long_key_name_that_exceeds_the_sixty_three_character_l_2": 11,
+    }
+
+    result = normalize_and_snakecase_keys(input_dict, special_case_keys)
+
+    assert result == expected_output, f"Expected {expected_output}, but got {result}"
