+++ f/connectors/comapeo/tests/comapeo_observations_test.py
@@ -2,46 +2,9 @@ import psycopg2
 
 from f.connectors.comapeo.comapeo_observations import (
     main,
-    normalize_and_snakecase_keys,
     transform_comapeo_observations,
 )
 from f.connectors.comapeo.tests.assets.server_responses import SAMPLE_OBSERVATIONS
-
-
-def test_normalize_and_snakecase_keys():
-    input_dict = {
-        "primaryKey": 1,
-        "camelCaseKey": 2,
-        "anotherCamelCaseKey": 3,
-        "keyWith-Collision": 4,
-        "keyWithCollision": 5,
-        "KeyWithCollision": 6,
-        "key-with-collision": 7,
-        "key_with_collision": 8,
-        "key_with_collision_2": 9,
-        "aVeryLongKeyNameThatExceedsTheSixtyThreeCharacterLimitAndNeedsTruncation": 10,
-        "aVeryLongKeyNameThatExceedsTheSixtyThreeCharacterLimitAndNeedsTruncationAlso": 11,
-    }
-
-    special_case_keys = set(["primaryKey"])
-
-    expected_output = {
-        "primaryKey": 1,
-        "camel_case_key": 2,
-        "another_camel_case_key": 3,
-        "key_with_collision": 4,
-        "key_with_collision_2": 5,
-        "key_with_collision_3": 6,
-        "key_with_collision_4": 7,
-        "key_with_collision_5": 8,
-        "key_with_collision_2_2": 9,
-        "a_very_long_key_name_that_exceeds_the_sixty_three_character_l_1": 10,
-        "a_very_long_key_name_that_exceeds_the_sixty_three_character_l_2": 11,
-    }
-
-    result = normalize_and_snakecase_keys(input_dict, special_case_keys)
-
-    assert result == expected_output, f"Expected {expected_output}, but got {result}"
 
 
 def test_transform_comapeo_observations():
